+++ app.ts
@@ -1,71 +1,71 @@
-import path from 'path'
-import * as Sentry from '@sentry/node'
-import bodyParser from 'body-parser'
-import busboy from 'connect-busboy'
-import cookieParser from 'cookie-parser'
-import cors from 'cors'
-import express, { Request } from 'express'
-import expressWs from '@small-tech/express-ws'
-import logger from 'morgan'
-import { NODE_ENV, sentryDsn, sessionSecret } from './config'
-import router from './router'
+import path from 'path';
+import * as Sentry from '@sentry/node';
+import bodyParser from 'body-parser';
+import busboy from 'connect-busboy';
+import cookieParser from 'cookie-parser';
+import cors from 'cors';
+import express, { Request } from 'express';
+import expressWs from '@small-tech/express-ws';
+import logger from 'morgan';
+import config from './config';
+import router from './router';
 
 interface LoadedRequest extends Request {
-  user: {}
+  user: {};
 }
 
 // Set up Sentry error tracking
 Sentry.init({
-  dsn: sentryDsn,
-  environment: NODE_ENV
-})
+  dsn: config.sentryDsn,
+  environment: config.NODE_ENV
+});
 
 // Express App
-const app = express()
+const app = express();
 
 /**
  * Account for nginx proxy when getting client's IP address
  * http://expressjs.com/en/guide/behind-proxies.html
  */
-app.set('trust proxy', true)
+app.set('trust proxy', true);
 
 // Setup middleware
-app.use(Sentry.Handlers.requestHandler()) // The Sentry request handler must be the first middleware on the app
-app.use(logger('dev'))
-app.use(bodyParser.json())
-app.use(bodyParser.urlencoded({ extended: true }))
-app.use(cookieParser(sessionSecret))
-app.use(express.static(path.join(__dirname, 'dist')))
-app.use(busboy())
+app.use(Sentry.Handlers.requestHandler()); // The Sentry request handler must be the first middleware on the app
+app.use(logger('dev'));
+app.use(bodyParser.json());
+app.use(bodyParser.urlencoded({ extended: true }));
+app.use(cookieParser(config.sessionSecret));
+app.use(express.static(path.join(__dirname, 'dist')));
+app.use(busboy());
 app.use(
   cors({
     origin: true,
     credentials: true,
-    exposedHeaders: NODE_ENV === 'dev' ? ['Date'] : undefined
+    exposedHeaders: config.NODE_ENV === 'dev' ? ['Date'] : undefined
   })
-)
+);
 // see https://stackoverflow.com/questions/51023943/nodejs-getting-username-of-logged-in-user-within-route
 app.use((req: LoadedRequest, res, next) => {
-  res.locals.user = req.user || null
-  next()
-})
+  res.locals.user = req.user || null;
+  next();
+});
 
 // The error handler must be before any other error middleware and after all controllers
-app.use(Sentry.Handlers.errorHandler())
+app.use(Sentry.Handlers.errorHandler());
 
 // Send error responses to API requests after they are passed to Sentry
 app.use(
   ['/api', '/auth', '/contact', '/school', '/twiml', '/whiteboard'],
   (err, req, res, next) => {
-    res.status(err.httpStatus || 500).json({ err: err.message || err })
-    next()
+    res.status(err.httpStatus || 500).json({ err: err.message || err });
+    next();
   }
-)
+);
 
 // initialize Express WebSockets
-expressWs(app)
+expressWs(app);
 
 // Load server router
-router(app)
+router(app);
 
-export { app }
+export default app;
